refactor: ViewerContent.tsx 리팩토링 (#125)
